Determine ENDPOINTS_SERVICE_VERSION if not specified
In line with the java library if a user specifies ENDPOINTS_SERVICE_NAME
but not ENDPOINTS_SERVICE_VERSION we should assume that they want the
most recent version. This greatly simplifies the automatic deployment
scenario where we generate the API specification via endpoints framework
library and then immediately deploy the code.

Closes: #68